stacker: initial commit
Initial commit of stacker, something to get started with.  It
only does the equivalent of umoci ls right now, and some basic
config.yml parsing.

We'll probably want to start using a real cli parser at some
point, but generally I don't like any of them, so prefer to do
it by hand as long as it's reasonable.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>